requirements.txt
#tensorflow
tensorflow-gpu==2.5.0

tqdm
pillow
numpy
scikit-learn
streamlit